refactor: fixture 분리
- 테스트에 사용하는 fixture 패키지 분리
