+++ src/test/java/com/atwoz/interview/domain/memberinterview/MemberInterviewTest.java
@@ -2,14 +2,14 @@ package com.atwoz.interview.domain.memberinterview;
 
 import com.atwoz.interview.domain.interview.Interview;
 import com.atwoz.interview.domain.interview.vo.InterviewType;
-import com.atwoz.interview.fixture.InterviewFixture;
+import com.atwoz.interview.fixture.interview.InterviewFixture;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
-import static com.atwoz.interview.fixture.InterviewFixture.인터뷰_나_일반_질문;
-import static com.atwoz.interview.fixture.InterviewFixture.인터뷰_나_일반_질문_id;
+import static com.atwoz.interview.fixture.interview.InterviewFixture.인터뷰_나_일반_질문;
+import static com.atwoz.interview.fixture.interview.InterviewFixture.인터뷰_나_일반_질문_id;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.SoftAssertions.assertSoftly;
